requirements.txt
absl-py==1.4.0
apturl==0.5.2
astunparse==1.6.3
blinker==1.8.2
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
filelock==3.6.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.40.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gyp==0.1
h5py==3.8.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jax==0.4.11
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.3.1
keras==2.12.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
ml-dtypes==0.2.0
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.5.0
protobuf==4.23.2
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.4
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
six==1.16.0
sklearn==0.0.post7
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.2.1
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==3.0.4
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
